bcrypt: Move GnuTLS support to a new file.
